Proxy the harmgen-call by a bash script
The generation-process must be able to generate the table/interface
files even when the typechecker reports errors. In the first place, this
is necessary when bootstraping a software project. The usage of
non-existing interfaces prevents other hack-processes to run if the
typechecker is enabled. Fixes #5

This fix may lead to some incompatibilities...